+++ src/Strings/index.js
@@ -1,0 +1,11 @@
+import { capitalize } from "./capitalize.js";
+
+
+/**
+ * Functions utils for strings
+ * 
+ * @namespace Strings
+ */
+export {
+	capitalize
+};